改了很多东西（？
加了形点的吸附效果
改了第二面弹幕的移动速度
改了combo有效时间
然后忘了（
（是不是忘了传修改过的关卡文件？

+++ SCRIPT/FUNC.py
@@ -11,7 +11,7 @@ def delta(minuend: tuple, subtrahend: tuple) -> tuple:
     return tuple(i - j for i, j in zip(minuend, subtrahend))
 
 
-def fibonacci(former: float, latter: float, frequency: int) -> float:
+def fibonacci(former: int, latter: int, frequency: int) -> int:
     """
     计算广义斐波那契数列的第frequency项 \n
     former和latter可以是浮点数
